--- a/openquake/qa_tests_data/scenario_damage/case_1c/expected/dmg_dist_total.xml
+++ b/openquake/qa_tests_data/scenario_damage/case_1c/expected/dmg_dist_total.xml
@@ -7,10 +7,10 @@ xmlns:gml="http://www.opengis.net/gml"
         <damageStates>
             no_damage slight moderate extreme complete
         </damageStates>
-        <damage ds="no_damage" mean="4.90665E-01" stddev="4.01763E-01"/>
-        <damage ds="slight" mean="3.24988E-01" stddev="2.54349E-01"/>
-        <damage ds="moderate" mean="7.08492E-02" stddev="9.01818E-02"/>
-        <damage ds="extreme" mean="2.11334E-02" stddev="3.17312E-02"/>
-        <damage ds="complete" mean="9.23640E-02" stddev="1.30216E-01"/>
+        <damage ds="no_damage" mean="4.79999E-01" stddev="3.86224E-01"/>
+        <damage ds="slight" mean="3.53799E-01" stddev="2.53865E-01"/>
+        <damage ds="moderate" mean="6.55346E-02" stddev="8.33401E-02"/>
+        <damage ds="extreme" mean="1.84490E-02" stddev="2.71265E-02"/>
+        <damage ds="complete" mean="8.22188E-02" stddev="1.10276E-01"/>
     </totalDmgDist>
 </nrml>