… openquake/qa_tests_data/scenario_risk/case_5/expected/losses_by_asset.csv