Companion of https://github.com/gem/oq-hazardlib/pull/543. The change in hazardlib has substantial impact on the performance. Consider for instance a case with 1 intensity measure type, 2560 sites and 1 million of ruptures; suppose than in average only half of the sites are affected by the ruptures (i.e. around 1280 sites per rupture are affected, not always the same).

*before the change* we would have computed 1 million of small matrices of size 1280 x 1280 corresponding to a memory allocation 1,000,000 * 1280 * 1280 * 8 bytes = 12 TB (terabytes!) which also has to be deallocated

*after the change* we will compute a single big matrix of size 2560 * 2560 corresponding to a memory
allocation of 50 MB, i.e. 240,000 times less memory.

In reality memory is allocated not only in the correlation matrix, but still the change in hazardlib guarantees a substantial speedup. For instance, the runtime a real life computation for Okhlahama (from one of ours users) with 2,684 sites and 60,761 ruptures went down from 27 hours to 1 hour (the speedup in building hazard was 44x).

Notice that the numbers are different because generating a single big matrix is different than generating multiple small matrices. The numbers are the same if the site collection is not filtered, each rupture affects all sites, i.e. all small matrices are the same as the big matrix. The root of the difference is on the Cholesky decomposition; if L(C) if the Cholesky decomposition of the correlation matrix C and P is a projection operator extracting the small matrix from the big one, one has

```python
L(P(C))  != P(L(C))
```

The values of the GMFs are sensitive to the Cholesky decomposition and the projection operator.
For instance, if there are 3 sites but for a given rupture only two are relevant, one could have the following correlation matrices:

```python
C_big = 
              1.00000 0.80886 0.52920
              0.80886 1.00000 0.65425
              0.52920 0.65425 1.00000

C_small =
                 1.00000 0.65425
                 0.65425 1.00000

As you see, C_small = P(C_big) is the projection of C_big on the sites 2, 3. However,
there is no such relation for the Cholesky matrices:

L_big = 
             1.00000 0.00000 0.00000
             0.80886 0.58801 0.00000
             0.52920 0.38470 0.75628

L_small = 
                 1.00000 0.00000
                 0.65425 0.75628

P(L_big) =
                  0.58801 0.00000
                  0.38470 0.75628

             != L_small
```

It should also be noticed that the computation of correlated GMFs has some interesting nonlocal behavior: for instance, if you add a site which is outside of the maximum distance for all ruptures, the Cholesky decomposition is different and the GMFs are different, contrarily to the idea that that site has no effect. This is actually correct, since getting zero correlation to zero for distant sites was an artifact of
the maximum distance truncation. In other words, with the old approach, we were using a projected correlation matrix

```python
C_proj = 
              0.00000 0.00000 0.00000
              0.00000 1.00000 0.65425
              0.00000 0.65425 1.00000
```

and this caused having a different Cholesky decomposition for different ruptures. With the new approach there is a single correlation matrix and a single decomposition for all ruptures.

PS: while at it, I fixed the profiler display that was broken for Python 3.